💡 refactor: create dialogList constant to remove duplicate code
