+++ app/_/components/layout/membersOnly/type.ts
@@ -1,0 +1,4 @@
+export type OneListProps = {
+    label: string
+    onClick: () => void | Promise<void>
+}
